// packages/valnor/src/core/__tests__/codegen/one_sql.order-table.ts
/* File generated by Valnor. Do not edit. */
import * as lib from "valnor";
import * as udt from "./one_sql-enums.js";

export const Order = lib.newTable({
   name: "order",
   schema: "one_sql",
   types: <{ Insert: IOrderInsert, Update: IOrderUpdate }>{}, 
   pk: "order_pk", 
},{

   /**
    * order_id uuid default gen_random_uuid()
   */
   orderId: "order_id",

   /**
    * status order_status default 'created'::one_sql.order_status
   */
   status: "status",

   /**
    * created_at timestamptz
   */
   createdAt: "created_at",

   /**
    * modified_at timestamptz
   */
   modifiedAt: "modified_at",

   /**
    * account_id uuid
   */
   accountId: "account_id",
});

export type IOrderInsert = {
   orderId?: string;
   status?: udt.OrderStatusUdt;
   createdAt: Date;
   modifiedAt: Date;
   accountId: string;
}

export type IOrderUpdate = Partial<IOrderInsert>;

export type IOrderSelect = {
   readonly orderId: string
   readonly status: udt.OrderStatusUdt
   readonly createdAt: Date
   readonly modifiedAt: Date
   readonly accountId: string
}

export type IOrderJson = lib.JsonRow<IOrderSelect>;